Omit useScreenSize, move detourTable to new component
